requirements.txt
discord.py[voice]
praw
youtube_dl
sqlalchemy
pymysql
aiohttp
python-dotenv